+++ gwcelery/tasks/orchestrator.py
@@ -386,8 +386,9 @@ def preliminary_alert(event, superevent_id):
         app.conf['preliminary_alert_trials_factor'][event['group'].lower()]
     if not event['offline'] \
             and trials_factor * event['far'] <= \
-            app.conf['preliminary_alert_far_threshold'] \
-            and {'DQV', 'INJ'}.isdisjoint(gracedb.get_labels(superevent_id)):
+            app.conf[
+                'preliminary_alert_far_threshold'][event['group'].lower()] \
+            and 'DQV' not in gracedb.get_labels(superevent_id):
         canvas |= (
             _create_voevent.s(
                 superevent_id, 'preliminary',